# Conflicts:
#	src/main/java/io/excaliburfrc/robot/commands/auto/oneBall/ballFive.java
#	src/main/java/io/excaliburfrc/robot/commands/auto/twoBalls/fiveFour.java

+++ src/main/java/io/excaliburfrc/robot/RobotContainer.java
@@ -152,6 +152,6 @@ public class RobotContainer {
     return chooser.getSelected();
   }
   public int getInitBalls() {
-    return initialBallCounter.getSelected().intValue();
+    return initialBallCounter.getSelected();
   }
 }
